Updated the data without navigation

+++ src/features/cart/cartSlice.js
@@ -2,6 +2,16 @@ import { createSlice } from "@reduxjs/toolkit";
 
 const initialState = {
   cart: [],
+
+  // cart: [
+  //   {
+  //     pizzaId: 12,
+  //     name: 'Mediterranean',
+  //     quantity: 2,
+  //     unitPrice: 16,
+  //     totalPrice: 32,
+  //   },
+  // ],
 };
 
 const cartSlice = createSlice({
@@ -9,28 +19,30 @@ const cartSlice = createSlice({
   initialState,
   reducers: {
     addItem(state, action) {
+      // payload = newItem
       state.cart.push(action.payload);
     },
     deleteItem(state, action) {
-      // payload === pizzaID
-      state.cart = state.cart.filter((item) => item.id !== action.payload);
+      // payload = pizzaId
+      state.cart = state.cart.filter((item) => item.pizzaId !== action.payload);
     },
     increaseItemQuantity(state, action) {
+      // payload = pizzaId
       const item = state.cart.find((item) => item.pizzaId === action.payload);
 
       item.quantity++;
       item.totalPrice = item.quantity * item.unitPrice;
     },
     decreaseItemQuantity(state, action) {
+      // payload = pizzaId
       const item = state.cart.find((item) => item.pizzaId === action.payload);
 
       item.quantity--;
       item.totalPrice = item.quantity * item.unitPrice;
 
-      if (!item.quantity === 0)
-        cartSlice.caseReducers.deleteItem(state, action);
+      if (item.quantity === 0) cartSlice.caseReducers.deleteItem(state, action);
     },
-    clearCart(state, action) {
+    clearCart(state) {
       state.cart = [];
     },
   },
@@ -48,51 +60,13 @@ export default cartSlice.reducer;
 
 export const getCart = (state) => state.cart.cart;
 
-export const getCurrentQuantityById = (id) =>
+export const getTotalCartQuantity = (state) =>
+  state.cart.cart.reduce((sum, item) => sum + item.quantity, 0);
+
+export const getTotalCartPrice = (state) =>
+  state.cart.cart.reduce((sum, item) => sum + item.totalPrice, 0);
+
+export const getCurrentQuantityById = (id) => (state) =>
   state.cart.cart.find((item) => item.pizzaId === id)?.quantity ?? 0;
 
-export const getTotalCartQuantity = (state) => {
-  state.cart.cart.reduce((sum, cur) => sum + cur.quantity, 0);
-};
-
-export const getTotalCartPrice = (state) => {
-  state.cart.cart.reduce((sum, cur) => sum + item.totalPrice, 0);
-};
-
-// const initialState = {
-//   cart: [
-//     {
-//       pizzaId: 22,
-//       name: "damilolo",
-//       quantity: 2,
-//       unitPrice: 13,
-//       total: "26",
-//     },
-//   ],
-// };
-
-// const cartSlice = createSlice({
-//   name: "cart",
-//   initialState,
-//   reducers: {
-//     addItem(state, action) {
-//       state.cart.push(action.payload);
-//     },
-//     deleteItem(state, action) {
-//       state.cart = state.cart.filter((item) => item.pizzaId !== action.payload);
-//     },
-//     increaseItemQuantity(state, action) {
-//       const item = state.cart.find((item) => item.pizzaId === action.payload);
-//       item.quantity++;
-//       item.total = item.quantity * item.unitPrice;
-//     },
-//     decreaseItemQauntity(state, action) {
-//       const item = state.cart.find((item) => item.pizzaId === action.payload);
-//       item.quantity--;
-//       item.total = item.quantity * item.unitPrice;
-//     },
-//     clearItem(state, action) {
-//       state.cart = [];
-//     },
-//   },
-// });
+// 'reselect'
